# server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "A microservice to support and simplify our applications",
    "version": "1.0.0",
    "title": "HackerBay Microservice",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:4000",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "user",
      "description": "Handles the user requests"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "logs in a user into the platform",
        "description": "A user can login using their username and password",
        "operationId": "login",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "login parameters",
            "required": true,
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succesfully logged in"
          },
          "422": {
            "description": "Missing required parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/json-patch": {
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "apply patch to a json object",
        "description": "applies different patch operations to a json object and return the patched object",
        "operationId": "patch",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "patch parameters",
            "required": true,
            "schema": {
              "$ref": "#/definitions/patch"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "patch successful"
          },
          "400": {
            "description": "invalid patch parameters"
          },
          "401": {
            "description": "Unauthorized access"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/user/thumbnail": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "creates a thumbnail",
        "description": "downloads an image from a url and return a thumbnail of it",
        "operationId": "thumbnail",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "thumb parameters",
            "required": true,
            "schema": {
              "$ref": "#/definitions/thumbnail"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "thumbnail successfully created"
          },
          "400": {
            "description": "invalid imageUrl"
          },
          "401": {
            "description": "Unauthorized access"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "login": {
      "type": "object",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "patch": {
      "type": "object",
      "required": [
        "docs",
        "patch"
      ],
      "properties": {
        "docs": {
          "type": "object"
        },
        "patch": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/patchObj"
          }
        }
      }
    },
    "patchObj": {
      "type": "object",
      "required": [
        "op",
        "path"
      ],
      "properties": {
        "op": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "thumbnail": {
      "type": "object",
      "required": [
        "imageUrl"
      ],
      "properties": {
        "imageUrl": {
          "type": "string"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}